pyproject.toml
[project]
name = "houseme"
version = "0.1.0"
description = "Craigslist apartment hunter with a TUI, scam detection, and AI-drafted visit request emails."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.89.0",
    "beautifulsoup4>=4.11.2",
    "click>=8.3.2",
    "imagehash>=4.3.2",
    "pillow>=12.2.0",
    "requests>=2.33.1",
    "textual>=8.2.3",
]

[project.optional-dependencies]
cdp = ["playwright>=1.58.0"]
